[chore] dev mode refactor (#3138)
* use vite.watcher instead of cheap-watch

* remove some indirection

* remove some more indirection

* more simplifying

* simplify

* lint

* remove code that apparently doesnt do anything

* apparently no longer necessary

* typecheck

* tests passing

* move plugin into separate module

* remove cheap-watch

* update lockfile

* revert versions in create-svelte
