+++ packages/create-svelte/templates/default/package.template.json
@@ -8,8 +8,8 @@
 		"preview": "svelte-kit preview"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "workspace:*",
-		"@sveltejs/kit": "workspace:*",
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/kit": "next",
 		"svelte": "^3.44.0"
 	},
 	"type": "module",
