+++ packages/kit/package.json
@@ -11,7 +11,6 @@
 	"type": "module",
 	"dependencies": {
 		"@sveltejs/vite-plugin-svelte": "^1.0.0-next.32",
-		"cheap-watch": "^1.0.4",
 		"sade": "^1.7.4",
 		"vite": "^2.7.2"
 	},
